Changed long to int in C++ script

+++ app/src/main/cpp/native-lib.cpp
@@ -6,16 +6,19 @@ Java_com_graphic_main_Vector2_quickInvSqrt(JNIEnv *env, jobject thiz, jfloat l) 
      * This is the fast inverse square root algorithm.
      * Calculates the value of 1 / sqrt(l) for a given float l in a more efficient way
      * than calculating it straight forward.
-     * Source: https://en.wikipedia.org/wiki/Fast_inverse_square_root
+     *
+     * Sources:
+     * https://en.wikipedia.org/wiki/Fast_inverse_square_root
+     * https://www.youtube.com/watch?v=p8u_k2LIZyo
      */
 
-    long i;
+    int i;
     float x2, y;
     const float threehalfs = 1.5F;
 
     x2 = l * 0.5F;
     y  = l;
-    i  = * ( long * ) &y;                       // evil floating point bit level hacking
+    i  = * ( int * ) &y;                       // evil floating point bit level hacking
     i  = 0x5f3759df - ( i >> 1 );               // what the fuck?
     y  = * ( float * ) &i;
     y  = y * ( threehalfs - ( x2 * y * y ) );   // 1st iteration
